+++ cogs/settings.py
@@ -9,13 +9,14 @@ from utils.tibia import tibia_worlds
 
 SETTINGS = {
     "world": {"title": "🌐 World", "check": lambda ctx: ctx.guild.id not in config.lite_servers},
-    # "newschannel": {"title": "📰 News channel"},
-    # "eventschannel": {"title": "📣 Events channel"},
-    "levelschannel": {"title": "🌟☠ Levels & Deaths channel", "check":
+    "newschannel": {"title": "📰 News channel"},
+    "eventschannel": {"title": "📣 Events channel"},
+    "levelschannel": {"title": "🌟☠ Tracking channel", "check":
         lambda ctx: ctx.guild.id not in config.lite_servers},
     # "prefix": {"title": "❗ Prefix"},
-    # "welcomemessage": {"title": "💬 Welcome message"},
-    # "askchannel": {"title": "🤖 Command channel"},
+    # "welcome": {"title": "💬 Welcome message"},
+    "askchannel": {"title": "🤖 Command channel"},
+    "commandsonly": {"title": "🗑 Command channel - Delete other"},
 }
 
 
@@ -35,9 +36,9 @@ class Settings:
         embed = discord.Embed(title=f"{SETTINGS[ctx.command.name]['title']} - {ctx.me.display_name} settings",
                               description=ctx.command.short_doc, color=discord.Color.blurple())
         embed.add_field(name="📄 Current value", value=current_value, inline=False)
-        embed.add_field(name="☑ Accepted values", value=accepted_values, inline=False)
         embed.add_field(name="📝 Edit", inline=False,
                         value=f"`{ctx.prefix}{ctx.command.full_parent_name} {ctx.invoked_with} [{edit_params}]`")
+        embed.add_field(name="☑ Accepted values", value=accepted_values, inline=False)
         embed.set_footer(text=f'Use "{ctx.prefix}help {ctx.command.full_parent_name} {ctx.invoked_with}" for more info')
         await ctx.send(embed=embed)
 
@@ -82,7 +83,7 @@ class Settings:
             await ctx.message.delete()
             return
 
-        set_server_property("world", ctx.guild.id, world)
+        set_server_property(ctx.guild.id, "world", world)
         self.bot.reload_worlds()
         if world is None:
             await ctx.send(f"{ctx.tick(True)} This server is no longer tracking any world.")
@@ -90,50 +91,27 @@ class Settings:
             await ctx.send(f"{ctx.tick(True)} This server is now tracking **{world}**")
 
     @checks.is_admin()
-    @settings.command(name="levelschannel", aliases=["deathschannel"])
+    @settings.command(name="levelschannel", aliases=["deathschannel", "trackingchannel"])
     async def settings_levelschannel(self, ctx, channel: str=None):
         """Changes the channel where levelup and deaths are announced.
 
         This is were all level ups and deaths of registered characters will be announced.
-        By default, I will use the highest channel on the list where I can send messages.
-        If the channel you assign is deleted, I will go back to using the top channel."""
-        current_channel_id = get_server_property("levels_channel", ctx.guild.id, is_int=True)
-        current_channel = ctx.guild.get_channel(current_channel_id)
-        print(current_channel_id)
+        By default, the highest channel on the list where the bot can send messages will be used.
+        If the assigned channel is deleted or forbidden, the top channel will be used again.
+
+        If this is disabled, Announcements won't be made, but there will still be tracking.
+        """
+        current_channel_id = get_server_property(ctx.guild.id, "levels_channel", is_int=True)
         if channel is None:
-            top_channel = self.bot.get_top_channel(ctx.guild, True)
-            if current_channel:
-                perms = current_channel.permissions_for(ctx.me)
-            else:
-                perms = discord.Permissions()
-            ok = False
-            if current_channel_id is None:
-                current_value = "None."
-            elif current_channel is None:
-                current_value = "None, previous channel was deleted."
-            elif not perms.read_messages or not perms.send_messages:
-                current_value = f"{current_channel.mention}, but I can't use the channel."
-            else:
-                current_value = f"{current_channel.mention}"
-                ok = True
-
-            if not ok:
-                # This condition should be impossible to meet, because if the bot can't send messages on any channel,
-                # it wouldn't be able to reply to this command in the first place ¯\_(ツ)_/¯
-                if top_channel is None:
-                    current_value += " I have no channel to announce."
-                else:
-                    current_value += f" I will use {top_channel.mention} meanwhile."
-
-            await self.show_info_embed(ctx, current_value, "A channel's name or ID.", "channel/none")
+            current_value = self.get_current_channel(ctx, current_channel_id)
+            await self.show_info_embed(ctx, current_value, "A channel's name or ID, or `disable`.", "channel/disable")
             return
-        if channel.lower() == "none":
-            if current_channel_id is None:
-                await ctx.send("There wasn't any levels and deaths channel set.")
+        if channel.lower() == "disable":
+            if current_channel_id is 0:
+                await ctx.send("Level and deaths announcements are already disabled.")
                 return
-            message = await ctx.send(f"Are you sure you want to remove <#{current_channel_id}> "
-                                     f"as the level & deaths channel?")
-            new_value = None
+            message = await ctx.send(f"Are you sure you want to disable the level & deaths channel?")
+            new_value = 0
         else:
             try:
                 new_channel = await commands.TextChannelConverter().convert(ctx, channel)
@@ -151,12 +129,218 @@ class Settings:
             await ctx.message.delete()
             return
 
-        set_server_property("levels_channel", ctx.guild.id, new_value)
-        self.bot.reload_worlds()
-        if new_value is None:
+        set_server_property(ctx.guild.id, "levels_channel", new_value)
+        if new_value is 0:
             await ctx.send(f"{ctx.tick(True)} The level & deaths channel has been disabled.")
         else:
             await ctx.send(f"{ctx.tick(True)} <#{new_value}> will now be used.")
+
+    @checks.is_admin()
+    @settings.command(name="newschannel")
+    async def settings_newschannel(self, ctx, channel: str=None):
+        """Changes the channel where Tibia news are announced.
+
+        This is where all news and articles posted in Tibia.com will be announced..
+        By default, the highest channel on the list where the bot can send messages will be used.
+        If the assigned channel is deleted or forbidden, the top channel will be used again.
+        """
+        current_channel_id = get_server_property(ctx.guild.id, "news_channel", is_int=True)
+        if channel is None:
+            current_value = self.get_current_channel(ctx, current_channel_id)
+            await self.show_info_embed(ctx, current_value, "A channel's name or ID, or `disable`.", "channel/disable")
+            return
+        if channel.lower() == "disable":
+            if current_channel_id is 0:
+                await ctx.send("News announcements are already disabled.")
+                return
+            message = await ctx.send(f"Are you sure you want to disable news announcements?")
+            new_value = 0
+        else:
+            try:
+                new_channel = await commands.TextChannelConverter().convert(ctx, channel)
+            except commands.BadArgument:
+                await ctx.send("I couldn't find that channel, are you sure it exists?")
+                return
+            perms = new_channel.permissions_for(ctx.me)
+            if not perms.read_messages or not perms.send_messages:
+                await ctx.send(f"I don't have permission to use {new_channel.mention}.")
+                return
+            message = await ctx.send(f"Are you sure you want {new_channel.mention} as the new news channel?")
+            new_value = new_channel.id
+        confirm = await ctx.react_confirm(message, timeout=60, delete_after=True)
+        if not confirm:
+            await ctx.message.delete()
+            return
+
+        set_server_property(ctx.guild.id, "news_channel", new_value)
+        if new_value is 0:
+            await ctx.send(f"{ctx.tick(True)} The news channel has been disabled.")
+        else:
+            await ctx.send(f"{ctx.tick(True)} <#{new_value}> will now be used for Tibia news.")
+
+    @checks.is_admin()
+    @settings.command(name="eventschannel")
+    async def settings_eventschannel(self, ctx, channel: str=None):
+        """Changes the channel where upcoming events are announced.
+
+        This is where announcements of events about to happen will be made.
+        By default, the highest channel on the list where the bot can send messages will be used.
+        If the assigned channel is deleted or forbidden, the top channel will be used again.
+
+        If this is disabled, users that subscribed to the event will stil receive notifications via PM.
+        """
+        current_channel_id = get_server_property(ctx.guild.id, "events_channel", is_int=True)
+        if channel is None:
+            current_value = self.get_current_channel(ctx, current_channel_id)
+            await self.show_info_embed(ctx, current_value, "A channel's name or ID, or `disable`.", "channel/disable")
+            return
+        if channel.lower() == "disable":
+            if current_channel_id is 0:
+                await ctx.send("Event announcements are already disabled.")
+                return
+            message = await ctx.send(f"Are you sure you want to disable events announcements?")
+            new_value = 0
+        else:
+            try:
+                new_channel = await commands.TextChannelConverter().convert(ctx, channel)
+            except commands.BadArgument:
+                await ctx.send("I couldn't find that channel, are you sure it exists?")
+                return
+            perms = new_channel.permissions_for(ctx.me)
+            if not perms.read_messages or not perms.send_messages:
+                await ctx.send(f"I don't have permission to use {new_channel.mention}.")
+                return
+            message = await ctx.send(f"Are you sure you want {new_channel.mention} as the new events channel?")
+            new_value = new_channel.id
+        confirm = await ctx.react_confirm(message, timeout=60, delete_after=True)
+        if not confirm:
+            await ctx.message.delete()
+            return
+
+        set_server_property(ctx.guild.id, "events_channel", new_value)
+        if new_value is 0:
+            await ctx.send(f"{ctx.tick(True)} The events channel has been disabled.")
+        else:
+            await ctx.send(f"{ctx.tick(True)} <#{new_value}> will now be used for events.")
+
+    @checks.is_admin()
+    @settings.command(name="askchannel", aliases=["commandchannel"])
+    async def settings_askchannel(self, ctx, channel: str=None):
+        """Changes the channel where longer replies for commands are given.
+
+        In this channel, pagination commands show more entries at once and command replies in general are longer."""
+        current_channel_id = get_server_property(ctx.guild.id, "ask_channel", is_int=True)
+        current_channel = ctx.guild.get_channel(current_channel_id)
+        if channel is None:
+            if current_channel:
+                perms = current_channel.permissions_for(ctx.me)
+            else:
+                perms = discord.Permissions()
+            ok = False
+            if current_channel_id is None:
+                current_value = f"None."
+            elif current_channel is None:
+                current_value = "Previous channel was deleted."
+            elif not perms.read_messages or not perms.send_messages:
+                current_value = f"{current_channel.mention}, but I can't use the channel."
+            else:
+                current_value = {current_channel.mention}
+                ok = True
+
+            if not ok:
+                current_value += f" By default, I'll use any channel named {config.ask_channel_name}."
+            await self.show_info_embed(ctx, current_value, "A channel's name or id, or `none`.", "channel/none")
+            return
+
+        if channel.lower() == "none":
+            if current_channel_id is None:
+                await ctx.send("There's no command channel set.")
+                return
+            message = await ctx.send(f"Are you sure you want to delete the set command channel?")
+            new_value = 0
+        else:
+            try:
+                new_channel = await commands.TextChannelConverter().convert(ctx, channel)
+            except commands.BadArgument:
+                await ctx.send("I couldn't find that channel, are you sure it exists?")
+                return
+            perms = new_channel.permissions_for(ctx.me)
+            if not perms.read_messages or not perms.send_messages:
+                await ctx.send(f"I don't have permission to use {new_channel.mention}.")
+                return
+            message = await ctx.send(f"Are you sure you want {new_channel.mention} as the new commands channel?")
+            new_value = new_channel.id
+        confirm = await ctx.react_confirm(message, timeout=60, delete_after=True)
+        if not confirm:
+            await ctx.message.delete()
+            return
+
+        set_server_property(ctx.guild.id, "ask_channel", new_value)
+        if new_value is 0:
+            await ctx.send(f"{ctx.tick(True)} The command channel was deleted."
+                           f"I will still use any channel named **{config.ask_channel_name}**.")
+        else:
+            await ctx.send(f"{ctx.tick(True)} <#{new_value}> will now be used as a command channel.")
+
+    @checks.is_admin()
+    @settings.command(name="commandsonly")
+    async def settings_commandsonly(self, ctx, option: str=None):
+        """Sets whether only commands are allowed in the command channel.
+
+        If this is enabled, everything that is not a message will be deleted from the command channel.
+        This allows the channel to be used exclusively for commands.
+
+        If the channel is shared with other command bots, this should be off.
+
+        Note that the bot needs `Manage Messages` permission to delete messages."""
+        def yes_no(choice: bool):
+            return "Yes" if choice else "No"
+        if option is None:
+            current = get_server_property(ctx.guild.id, "commandsonly", is_int=True)
+            if current is None:
+                current_value = f"{yes_no(config.ask_channel_delete)} (Global default)"
+            else:
+                current_value = yes_no(current)
+            await self.show_info_embed(ctx, current_value, "yes/no", "yes/no")
+            return
+        if option.lower() == "yes":
+            set_server_property(ctx.guild.id, "commandsonly", True)
+            await ctx.send(f"{ctx.tick(True)} I will delete non-commands in the command channel from now on.")
+        elif option.lower() == "no":
+            set_server_property(ctx.guild.id, "commandsonly", False)
+            await ctx.send(f"{ctx.tick(True)} I won't delete non-commands in the command channel from now on.")
+        else:
+            await ctx.send("That's not a valid option, try **yes** or **no**.")
+
+    def get_current_channel(self, ctx, current_channel_id):
+        top_channel = self.bot.get_top_channel(ctx.guild, True)
+        current_channel = ctx.guild.get_channel(current_channel_id)
+        if current_channel:
+            perms = current_channel.permissions_for(ctx.me)
+        else:
+            perms = discord.Permissions()
+        ok = False
+        if current_channel_id == 0:
+            current_value = "Disabled."
+            ok = True
+        elif current_channel_id is None:
+            current_value = "None,"
+        elif current_channel is None:
+            current_value = "None, previous channel was deleted."
+        elif not perms.read_messages or not perms.send_messages:
+            current_value = f"{current_channel.mention}, but I can't use the channel."
+        else:
+            current_value = f"{current_channel.mention}"
+            ok = True
+
+        if not ok:
+            # This condition should be impossible to meet, because if the bot can't send messages on any channel,
+            # it wouldn't be able to reply to this command in the first place ¯\_(ツ)_/¯
+            if top_channel is None:
+                current_value += " I have no channel to use."
+            else:
+                current_value += f" I will use {top_channel.mention} meanwhile."
+        return current_value
 
 
 def setup(bot):