+++ utils/context.py
@@ -3,13 +3,24 @@ import asyncio
 import discord
 from discord.ext import commands
 
+from utils.config import config
+from utils.database import get_server_property
+
 YES_NO_REACTIONS = ('\U0001f1fe', '\U0001f1f3')
 CHECK_REACTIONS = ('\N{WHITE HEAVY CHECK MARK}', '\N{CROSS MARK}')
 
 
-class Context(commands.Context):
+class NabCtx(commands.Context):
     def __init__(self, **kwargs):
         super().__init__(**kwargs)
+
+    @staticmethod
+    def tick(value: bool, label: str = None):
+        """Displays a checkmark or a cross depending on the value."""
+        emoji = CHECK_REACTIONS[int(not value)]
+        if label:
+            return emoji + label
+        return emoji
 
     @property
     def world(self):
@@ -18,13 +29,32 @@ class Context(commands.Context):
         else:
             return self.bot.tracked_worlds.get(self.guild.id, None)
 
-    @staticmethod
-    def tick(value: bool, label: str=None):
-        """Displays a checkmark or a cross depending on the value."""
-        emoji = CHECK_REACTIONS[int(not value)]
-        if label:
-            return emoji+label
-        return emoji
+    @property
+    def long(self):
+        """Whether the current context allows long replies or not
+
+        Private messages and command channels allow long replies.
+        """
+        if self.guild is None:
+            return True
+        return self.is_askchannel
+
+    @property
+    def is_askchannel(self):
+        """Checks if the current channel is the command channel"""
+        ask_channel_id = get_server_property(self.guild.id, "ask_channel", is_int=True)
+        ask_channel = self.guild.get_channel(ask_channel_id)
+        if ask_channel is None:
+            return self.channel.name == config.ask_channel_name
+        return ask_channel == self.channel
+
+    @property
+    def ask_channel_name(self):
+        ask_channel_id = get_server_property(self.guild.id, "ask_channel", is_int=True)
+        ask_channel = self.guild.get_channel(ask_channel_id)
+        if ask_channel is None:
+            return config.ask_channel_name
+        return ask_channel.name
 
     async def react_confirm(self, message: discord.Message, *, timeout=120.0, delete_after=False,
                             use_checkmark=False):